feat: refactor imports to use helpers; add date and item-data helpers; implement JWT signing and verification; add scroll-lock helpers; introduce upload components
